:building_construction: Improve React jsx code architecture
