+++ frontend/src/components/BottomNav.jsx
@@ -1,0 +1,22 @@
+import { BottomNavigation, BottomNavigationAction } from "@mui/material";
+import { EmojiEvents, HowToVote, Info, ViewStream } from "@mui/icons-material";
+
+export default function BottomNav({ state: [tab, setTab] }) {
+  return (
+    <BottomNavigation
+      className="BottomNav"
+      showLabels
+      value={tab}
+      onChange={(event, newValue) => setTab(newValue)}
+    >
+      {[
+        { key: "votes", label: "Voter", icon: <HowToVote /> },
+        { key: "resultats", label: "Résultats", icon: <EmojiEvents /> },
+        { key: "mes-votes", label: "Mes Votes", icon: <ViewStream /> },
+        { key: "a-propos", label: "À Propos", icon: <Info /> },
+      ].map(({ label, icon, key }) => (
+        <BottomNavigationAction label={label} icon={icon} key={key} />
+      ))}
+    </BottomNavigation>
+  );
+}
