+++ frontend/src/data/variables.jsx
@@ -1,0 +1,2 @@
+export const minVotes = 5;
+export const projectURL = "https://github.com/arnaudsm/votefinder.fr";
