+++ frontend/src/utils/utils.jsx
@@ -1,0 +1,20 @@
+export function formatDate(txt) {
+  if (!txt) return "";
+  try {
+    return new Intl.DateTimeFormat("fr-FR", {
+      dateStyle: "long",
+    }).format(new Date(txt));
+  } catch (error) {
+    console.log(error);
+  }
+  return txt;
+}
+
+export function shuffle(arr) {
+  const newArr = arr.slice();
+  for (let i = newArr.length - 1; i > 0; i--) {
+    const rand = Math.floor(Math.random() * (i + 1));
+    [newArr[i], newArr[rand]] = [newArr[rand], newArr[i]];
+  }
+  return newArr;
+}
